requirements.txt
Flask==1.1.1
Pillow==4.2.1
opencv-python==4.1.1.26
pytesseract==0.3.0
pymongo==3.9.0
cloudinary==1.19.1
virtualenv==16.7.7
requests==2.22.0
requests-oauthlib==1.3.0